Refactor to use dataloader.load() only
